data reconstruction from sparse modes
This is running but not producing sensible output. When I reconstruct
using all of the modes I end up unable to get anything that looks like
the original data, even in the mean.

This is possibly a limitation caused by the length of my dataset (too
short), or maybe because I've multiplied matrices together incorrectly.

To try: can we reconstruct from the modes and optimal amplitudes,
without using the sparse calculations? i.e, U*Ydmd, Vand, xdmd